Add CLI command test setup
Implement tests for some CLI commands. I started with these tests as
they test the whole chain and not individual methods of classes. When I
was working on this prototype I tested most things manually so I want to
test the whole thing first before testing individual parts.

I realized I need to call certain commands to use their output
artifacts, like `gem build` for the `*.gem` files. This requires me to
actually fetch dependencies if necessary, like for Elixir. This
increases the test suite time by a lot, so perhaps these tests need to
be extracted later to an integration test suite. Maybe even run in a
Docker container so you don't need to have all the supported languages
installed locally. And so we can more closely replicate the test
environment.

I don't just want a test suite that asserts if certain commands are
called. This method of testing is how I found out already there was a
problem with the Ruby gem publish order and I split out the Git commit
and push steps.

I want to test the whole chain interacts with each other correctly, with
the exception of the publishing to the package manager register and git
remote repository. So we don't accidentally publish anything during a
test run.