Add VersionPromoter to bump version numbers
Add class to help promote/bump version numbers. Extract this from the
Package class so that it can be used (and tested) separately.

I added a lot of tests. This makes sure all scenarios are covered, all
the different version bumps and all the different scenarios that arise
with prerelease versions.

It did not implement support for prerelease downgrades: It's not
possible to go from 1.0.0-beta.1 to 1.0.0-alpha.2. For this I would need
to add parsing for existing git tags to see what the latest release for
the downgraded prerelease type is and that's too much work right now. I
also don't know if there's a situation where you would want to downgrade
the prerelease that's not just another prerelease version of the same
type.