Only publish Ruby .gem files for published version
The Ruby publishing process doesn't clean up pushed/published .gem
files. So make sure we only push the versions of the to-be-published
version, and not older .gem version files as well.

Remove the required `publish.gem_files` option by supporting it for
"normal" publishing in the root directory as well.

Rename the option to be more specific that it's about the parent dir,
and not the files themselves. Mono will look for the right version .gem
files itself.