Fix mono publish confirmation exit type
It would raise an error, and be caught by the Mono error handling, while
it should instead exit with a status code cleanly.

For the user the behavior is the same, it exits, but it shouldn't print
a message that indicates something went wrong.

Before:
```
Do you want to publish the above changes? (Y/n) n
A Mono error was encountered during the `mono publish` command. Stopping operation.

Mono::Error: 1
```

After:

```
Do you want to publish the above changes? (Y/n) n
Exiting...
```