Improve support for Node.js (npm/yarn) workspaces
## The problem

The original implementation performed the commands in the directories of
the workspaces/packages themselves, but this could lead to missing
dependencies that were shared in the project's root `package.json` file.

## Solution

Instead run commands via `npm <command> --workspace=package_name` so
that this problem does not occur.
The same for yarn, but with slightly different syntax:
`yarn workspace package_name <command>`.

Sources:
- https://docs.npmjs.com/cli/v7/using-npm/workspaces
- https://classic.yarnpkg.com/en/docs/cli/workspace

## Node.js package names

The problem I found was that I could not use the package directory name,
but had to use the name of the package specified in the `package.json`
file. We need to read the `package.json` file, parse the JSON and fetch
the `name` configuration to be able to run commands in
workspaces/packages of which the directory name does not match the
package name.

```
# packages/apollo-server/package.json
{
  "name": "@appsignal/apollo-server"
}
```

```
# Does not work
npm <command> --workspace=apollo-server
# Does work
npm <command> --workspace=@appsignal/apollo-server
```

## Implementation details

To make this work for all languages I had to change how commands are
called. Ruby and Elixir need to run the commands in the package
directories themselves, but Node.js shouldn't.

Instead of wrapping all commands in the `Package#chdir` block (to run
them in the package dir), this decision is now part of the language
specific Package class. So Node.js doesn't, but Ruby and Elixir do wrap
the command in `Package#chdir`.