Skip script in packages without script configured
In Node.js packages scripts don't have to be configured, such as build
and test.
When they are not available, skip the script for that package, but print
a message about that the script was skipped for that package. This
replicates the behavior I see in lerna where it silently skips commands
in a similar way.

This should bring us one step closer to using mono in the CI for
Node.js integration mono repo.