Support different version.rb paths
For the Ruby packages, search for the first version.rb in the project,
and update that version file, instead of hardcoding it to
`lib/appsignal/version.rb`, that only works for the AppSignal Ruby gem.

Yes, there may be multiple version.rb files, but that's not a problem
yet. We can always fix that later, but I don't want to overthink it yet.

This allows us to hopefully release mono with mono later on.