Add language clean step to Node.js mono repos
Also make sure to clean the root `node_modules` directory that is shared
between workspaces in Node.js. It now does the same clean operation
twice for single package projects, but that's fine for now. I could skip
it for mono repos, but if users customize the commands we may skip
commands. Since this doesn't break anything, let's keep it for now.

- No extra clean up needed for Ruby and Elixir.
- Update the tests to have the correct naming.