Validate `--package` names
When specifying a package with the `--package` CLI option, mono would
not alert the user that the command was not run if the package could not
be found. It would silently exit.

This change will make mono exit with an error if no package with the
specified name does not exist to communicate what happens to the user
more clearly.

Fixes #8