Make run_command_in_package thread safe
I want to add a --parallel option, but using `Dir.chdir` with a block is
not thread safe and will raise an error. The `Dir.chdir` behavior is
process specific on the OS-level, and cannot be thread specific, so we
need to change how we navigate to directories for commands.

Instead pass the path to the Command class and have it figure out the
location in which to execute the command. Right now it only prefixes the
command with a `cd path/to/package && ...`. This option does allow for
more flexibility if we ever move away from the naive approach using
`system` to run commands.