Add --parallel option to mono run
Allow users to run commands in parallel. Start a new thread per command
and run those commands in parallel.

Testing is a naive approach. It doesn't actually test if the commands
are run in parallel, if the threads are running at the same time. It
only checks if the commands are run at all and doesn't cause any errors.

Ideally mono is run in the test suite as its own executable. Then we can
check if both commands are running at the same time with `ps aux` and
running a long running command like `tail -f /dev/null`. Something to
improve later.

I haven't added --parallel to other commands for now. We can add them
later if necessary.