Add unbootstrap command, redesign clean command (#9)
The "unbootstrap" command will have the same behavior as the "clean"
command did previously.

The "clean" command now will be focussed on cleaning the project back to
the state before the "build" command. For Ruby this means removing the
`*.gem` files. For Elixir this means removing the `_build` directory.
And for Node.js this means running the `clean` script configured in
`package.json`.

This split in behavior allows us to be more specific in what actions we
want to perform. We usually don't want to unbootstrap a project, but
clean it instead, so we can run another fresh build.