Update dependent packages on update
Updates packages automatically when a package they depend on in this
project also gets updated. Only works for Node.js and Ruby for now.
Elixir has a process issue.

## Scenario

Given a project has the packages: A, B and C

- Package A has no dependencies
- Package B depends on package A
- Package C depends on package B

When package A updates, we want all packages that depend on package A to
also update. This is achieved with a strict version lock for now, as
it's easiest to implement.

Later on we also want to support less strict version lock, not on patch
level but on minor releases as well, e.g.  `~> 1.2.0` that matches all
release in the 1.2 series.

When package A gets updated, package B also gets a patch level bump.
Also package C will be updated as it depends on package B, which is
updated because package A is updated.

When a package would not get a release on its own, it does get a new
release when a dependency gets updated.

## Ruby mono repo support

This change also makes Ruby mono repo support actually work. Previously
there was no test for this scenario so there was no guarantee it worked.
This change adds specs for the Ruby mono repo scenario and fixes the
`gem push` command for mono repos.