Fix version promotion from prereleases
The VersionPromoter didn't support promoting versions that were a
prerelease to a base release.

What would happens is:
- If a library has version 1.2.3-alpha.1
- When `mono publish` is run (without a prerelease option)
- The command would error.

The problem was caused by the check if prerelease are larger than one
another, so that you can't release an alpha version when the current
version is a beta version.

This change skips the check, if the new prerelease is not present (and
the new version is base/final release).

This fix allows us to publish versions final releases after the
prerelease track.