Fix packages being published multiple times
The updated_packages variable would track packages multiple times if
they were updated by a dependency or because they had a changeset.

Fix this by using a `Set` (list with unique values) to store the list of
updated packages, so that every package would only be listed as needing
an update once.