Fix dependent package versions to prereleases
When a package is updated because a dependency is updated, and it's a
prerelease, write the correct prerelease version as the dependency
version lock.

For version package A `1.2.3-alpha.1` it would write the dependency in
package B as `1.2.3` (without the prerelease).

To fix this, mark all packages as a prerelease before we figure out
which packages will be changed. This is because
`Package#update_dependency` write the version number as a string on the
`@updated_dependencies` hash, so when the package gets marked later as a
prerelease the version number would not be updated later.