Support final releases from a prerelease
When a package has no changes (changesets), there may be a time where
it's good to release without making any other changes.

Support this scenario in `mono publish` without changes, by bumping the
version to the final version.

This is not supported for mono publish with a prerelease, because I
would expect some changes to publish for a new prerelease.