Omit Git SHA from MemoryChangesets
When a MemoryChangeset is created, it's usually for some change mono
itself made, like updating dependency versions or releasing the final
version of a package (when it's a prerelease).

When it creates changelog entries for these changes, the Git SHA points
to the last commit, any commit, which is inaccurate.

The commit in which this change is actually stored is not made yet, so
omit it from the changelog. This creates less noise in the changelog.