Implement dependency tree
When building the appsignal-javascript mono repo from scratch the build
would fail as it depends on packages in the project that are not yet
compiled. This was because it tried to compile packages in the
alphabetical order, rather than what packages depend on which other
packages.

This move and expansion of the dependency tree from the PackagePromoter
adds a `packages` method that returns the packages in the order of least
dependencies first. This should make sure package A is compiled before
package B.