Bump requirement for npm to 7.12
Locally everything works with Node.js 15, but on CI it uses npm 7.7 and
it fails on it finding a specific workspace. This may be a bug that's
fixed in a more recent version, so bump it to work around this problem.

```
$ mono build
...
npm ERR! No workspaces found:
npm ERR!   --workspace=@appsignal/nodejs
```