Add module attribute version option for testing
To allow testing with packages that _don't_ put the version in a
module attribute, this patch adds an option to the
`create_package_mix` method that switches module attribute versions on
or off.

With `:version_in_module_attribute` set to `true`:

    defmodule MyPackage.Mixfile do
      use Mix.Project

      @source_url "https://github.com/appsignal/test-package"
      @version "1.2.3"

      def project do
	[
	  app: :test_package,
	  version: @version,

With `:version_in_module_attribute` set to `false`:

    defmodule MyPackage.Mixfile do
      use Mix.Project

      @source_url "https://github.com/appsignal/test-package"

      def project do
	[
	  app: :test_package,
	  version: "1.2.3",

NOTE: This patch includes a `rubocop:disable
Style/OptionalBooleanParameter`, as the suggestion caused a deprecation
warning in Ruby because the argument has a default value. The fix for
that was too verbose to be feasable[^1].

[^1]: https://bloggie.io/@kinopyo/how-to-fix-ruby-2-7-warning-using-the-last-argument-as-keyword-parameters-is-deprecated