Handle non-module-attribute Elixir versions
Add a test for version numbers that are placed into the `mix.exs`
project block directly instead of being passed through the `@version`
module attribute, and update the regular expression to match on both.

NOTE: Since this patch, the version number comes out of the second
match in the expression instead of the first one. That's because the
first match is taken up by either `@version` or `version:`, as the
regulax expression includes `(@version|version:)`. Another option
would have been to make both the `@` and `:` optional (`@?version:?`),
but that would match invalid version specifications like "version
1.2.3".

The regular expression is further altered by adding a third match that
matches the comma if it exists, which is used to write the version
back to the file to make the publishing test pass.