Move Ruby specific CI setup to scripts
This way they're easier to reuse between tasks, and we can add other
jobs that do not require all this setup more easily without wasting time
waiting on the global prologue and epilogue.

It's not possible to use `set -u` for Semaphore scripts as they will
fail in undefined variable usage, so we can only enable it after calling
the `sem-version` scripts.