Fix yarn publish prompting user for version number
When calling `yarn publish` the command prompts the user to enter a new
version number. Mono already knows this version number, so pass it along
to the `yarn publish` command.

I wish yarn would not prompt this at all because we've updated the
package version by updating the `package.json` file. Passing in the
`--new-version` option seems to not do anything else than update the
`package.json` as well, so there should be no problem specifying it like
this. It won't do anything, which is what we want.