Add mono-dev executable
Add `mono-dev` executable to allow for testing mono. The `mono`
executable will not allow usage of mono with uncommitted changes. This
will prevent accidental usage of mono with uncommitted changes that
would cause mono to crash unexpectedly.

I've added another `mono-dev` executable that behaves the same way as
`mono` would previously. And added a check to `mono` to see if there are
any uncommitted changes and exit if there are any.

Closes #20