Retry publish command on failure
When publishing a package and the command would fail, mono would exit
entirely. This would happen if a user would enter an invalid OTP code
during the process for example.

When in a mono package repo this could exit the entire process, such as
like on package 3 of 10. The only way to publish the package properly
then (as part was already published) is to publish them manually.

This retry option will prompt users on failure of the command to retry
it.

I've only added it to the publish command for now because that's the
only problematic one I can think of right now. If the build step fails
nothing is in a state of being half published yet. Pushing the Git
branch and tags should also be easy to do manually.

Fixes #24