Fix mono production check directory
The Git commands were run in the directory mono was run in as well, not
necessarily the directory in which the mono Git repository is present.

Use chdir to temporarily run the command in mono installation directory.

I'm using spawn instead of the backticks to "chdir". I could have also
wrapped it in `Dir.chdir { ... }`, but it seemed better to only run the
command in that directory rather than "move" the entire program within
that block to the installation directory.