Add `fish` support in setup script
Add support for the `fish` shell in the `script/setup`
installation script.

In addition, modify the script so that it does not exit early
after setting up the `PATH` for any one shell; rather, it will
continue, setting the `PATH` for all the shells it can find.

Also, modify its behaviour when writing to `~/.bashrc` and
`~/.zshrc` so that it is idempotent, not writing a line that
modifies the `PATH` if it's already present in the file.