Print changeset summaries on publish
To make it easier to tell what's being published, print a summary for
every changeset in every package.

The changeset contents are printed as single lines (even if the actual
changeset file contents is multi line) and is trimmed to maximum 100
characters to not break the CLI output formatting.

The changesets are sorted by type of version bump, with "major" version
bumps listed at the top and patch releases at the bottom.

The additional output made it a bit more difficult to test, which is why
I added a `strip_changeset_output` helper that we can run the output
through to filter out the changeset output for tests that don't actually
test it. We don't need to test it every time, only the specific tests
that test this summary.

Closes #21