Add types to changesets
Support different types of changesets:
add, change, fix, deprecate, remove and security.

These different types help communicate to the reader what type of
release a package version is and what types of changes it included.

This adopts the same types of changes as the Keep a Changelog format
described here: https://keepachangelog.com/en/1.0.0/
But doesn't adopt it entirely.

The changeset metadata gets an additional `type` field, used to indicate
the type of change. The naming is not matching the label in the
changelog exactly. I think this short naming is easier to remember when
writing a changeset. People can also use the `mono changeset add` tool
and it will prompt them for the type.

The `ChangesetCollection#write_changesets_to_changelog` method could
probably use a refactor at this point. But my head is not in the right
place for that right now. Also, there previously were no unit tests for
it even, so let's start small.

Closes #31