Prevent duplicate dependency changesets (#44)
Track which dependencies are updated in a package, and only add another
changeset if it wasn't already updated. There should never be the
scenario that the same package is updated to another type of version
bump, so the message should always be the same.

I created a new `DependencyBumpMemoryChangeset` class to track the types
of in memory changesets that created during this process and to be able
to compare them.