Support package.json devdependencies on publish
When I published the JavaScript packages I noticed the
`@appsignal/types` package had to versions in `yarn.lock`.
This was because the package was specified in the `devdependencies` of
the `@appsignal/plugin-breadcrumbs-console` package. Mono only checked
the `dependencies` and `optionalDependencies`, so the package version
was not updated.

This was not really a problem because it's only used during development
and testing, but it's not good to not update the package version and
possibly keep testing against an older version.

This change will update any package in the workspace listed in
`devdependencies` upon `mono publish`.