Support custom languages
Add a "custom" language option toallow for customization of every
command and step in mono. This way you don't have to reuse another
language option and potentially running scripts not meant for that
language.

Allow custom package version scripts to be run so we don't need to
implement this for Python right now, while we figure out how we want it
to work.