Allow output capture of package commands
Switch from using backticks to run commands to using `spawn`. Spawn has
more options to customize the command. We no longer have to use
`cd #{path}`, but can specify the changing of directories with the
`:chdir` option.

When `:capture => true` is given as an option, it will capture and
return all the STDOUT and STDERR from the child process.