Improve publish commit message
When a single package is published, don't say "Publish packages", but
use the singular form and show the version number in the subject.

Add a small description of what was changed to pass the Lintje
MessagePresence rule. The version number alone isn't enough in most
cases for single package releases.