Implement Python version scheme
Split the existing Mono::Version type into three subtypes for Ruby,
Semver (Elixir/Node.js) and Python. Implement separate `self.parse`
and `.to_s` methods for these subtypes.

Internally, the Python version type stores its pre-release type in
the forms already used by Mono ("alpha"/"beta"/"rc") and not the
forms used by Python version strings ("a"/"b"/"rc"), converting
between the two at the parse/serialise boundaries. This allows for
existing code that expects the Mono forms to work correctly.

Implement a Python version subtype that conforms to PEP 440 to the
extent of the Mono feature set. Features not present in Mono, such
as post-releases and local builds, are not supported.

Add a "version_scheme" key to the config, with possible values of
"semver", "ruby" and "python", defaulting to "semver". Use this
version scheme, if present, with the custom package type, in order
to parse and serialise versions.