Add mono publish --no-git flag
To make it easier to publish our agent, add this option to mono publish
so that our publish script in the agent repository can continue doing
all the work in Git, and it can wrap mono instead.

I thought about adding a new CLI command just for the changelog
generation and package version updating. That requires a bunch of
refactoring of the CLI and adding more tests to also cover all the
scenarios in that new subcommand. This is a smaller change.

Part of #48