Format changelog to move link to end of line (#77)
Make generated changelogs more readable by moving he link and changeset
bump to the end of the line. Focus on the change, rather than the
technical details.

On multi line changes it will put the details at the end on a new line.

This matches the https://keepachangelog.com/en/1.1.0/ format better.