Add auto-update functionality
When launching `mono`, check for updates by attempting to fetch the
`origin` remote in the repository. If an update is found, reset the
repository to the state in the remote, and restart mono by replacing
the current process with a new `bin/mono` invocation.

Only attempt to check for updates if it has not been checked for
updates in the last 24 hours. Ensure that failure to check for updates
(e.g. no internet connection) does not cause Mono to crash.