Add `MONO_DEV` environment variable
Add a `MONO_DEV` environment variable that, when set, causes `mono`
to behave like `mono-dev` -- that is, it does not refuse to run if
the code has been modified, and it does not attempt to auto-update.
This is useful in scenarios where the invocation of `mono` under test
is itself performed by another tool, such as in the AppSignal agent's
release scripts.

Also move the code around to avoid having three levels of nesting,
making RuboCop and future readers happy.