Ignore commits that modify changesets
I pushed a release after I changed some changesets to fix typos and
improve the changeset copy. Now when someone links to the relevant
change in the changelog they are linked to that "improve changeset copy"
commit. That's not the best experience. A `[skip mono]` tag in a commit
message will have Mono ignore that commit and link to the previous
commit that changed/created the changeset file.