List multiple commits in changelog per entry
When I changed a changeset entry's text I inadvertently made the
changelog link to that commit, rather than the relevant change.
If there are multiple changes to the changeset, list them all in the
changelog.

This will help identify all related changes if the changeset was changed
multiple times, for example to fix a typo. Preferably it would also use
the "skip mono" tag introduced the previous commit, but this way no
changes are missed.