Fix bug for changesets without commit (#81)
I made a commit that added a _new_ changeset and used the "skip mono"
tag. Mono then raised error when it tried to fetch the commit's date
when there was none. Fix this by falling back on `nil`. The order
doesn't matter too much. The date the option fallback for the sort.